+++ discovery/gossiper.go
@@ -20,6 +20,7 @@ import (
 	"github.com/lightningnetwork/lnd/chainntnfs"
 	"github.com/lightningnetwork/lnd/channeldb"
 	"github.com/lightningnetwork/lnd/channeldb/models"
+	"github.com/lightningnetwork/lnd/graph"
 	"github.com/lightningnetwork/lnd/keychain"
 	"github.com/lightningnetwork/lnd/kvdb"
 	"github.com/lightningnetwork/lnd/lnpeer"
@@ -169,7 +170,7 @@ type Config struct {
 	// topology of lightning network. After incoming channel, node, channel
 	// updates announcements are validated they are sent to the router in
 	// order to be included in the LN graph.
-	Router routing.ChannelGraphSource
+	Router graph.ChannelGraphSource
 
 	// ChanSeries is an interfaces that provides access to a time series
 	// view of the current known channel graph. Each GossipSyncer enabled
